Transpile the CLI and combine src/bin and src/index
