+++ server/package.json
@@ -7,7 +7,7 @@
   "main": "./out/server.js",
   "typings": "./out/server.d.ts",
   "bin": {
-    "bash-language-server": "./bin/main.js"
+    "bash-language-server": "./out/cli.js"
   },
   "repository": {
     "type": "git",
